refactor(vault-persistence): address code review findings (#10)
ペテルギウス指摘:
- mapping.rs: vault_header_to_params/record_to_params を Result 返却に変更し
  unwrap_or_else("1970-01-01T00:00:00Z") を削除。Fail Fast を徹底
- audit.rs: 自由関数5個を struct Audit のメソッドに変更（公開関数禁止原則）
- repository.rs: load/save を load_inner/save_inner + select_vault_header/select_records
  プライベートメソッドに分割し audit ボイラープレート20回を排除
- repository.rs: elapsed_ms を impl SqliteVaultRepository の関連関数に移動
- repository.rs: bytes_written の .unwrap_or(0) を PersistenceError::Io で return に変更

ペガサス指摘:
- mapping.rs (822行) を mapping/mod.rs (本番コード) + mapping/tests.rs (テスト) に分割

服部平次指摘:
- deny.toml: MPL-2.0 を licenses allow に追加 (hashlink 0.9 / rusqlite 依存)
- deny.toml: thiserror/windows-*/rustix/linux-raw-sys/getrandom/hashbrown 0.17 の
  多重バージョン skip エントリを追加
- .cargo/config.toml を git 管理から除外し .gitignore に追加
- SqliteVaultRepository::with_dir を pub → pub(crate) に変更
