+++ .gitignore
@@ -1,0 +1,1 @@
+.cargo/config.toml
